requirements.txt
attrdict>=2.0.1
h5py>=2.10.0
imageio>=2.6.1
git+https://github.com/aleju/imgaug
numpy>=1.17.3
opencv-python==4.1.2.30
git+https://github.com/jamesbowman/openexrpython.git
oyaml>=0.9
pathlib>=1.0.1
Pillow>=6.2.0
scikit-image>=0.15.0
scipy>=1.3.1
Shapely>=1.6.4.post2
tensorboardX>=1.9
tensorflow>=1.14.0
termcolor>=1.1.0
torch>=1.3.0
torchvision>=0.4.1
tqdm>=4.36.1
open3d>=0.8.0.0
